refactor(common): widen public API and common layer to 64-bit moduli
Raise the shared NTT surface from uint32_t to uint64_t so prime moduli up to
2^63 are supported.

- ntt_is_prime() now runs deterministic Miller-Rabin over the full
  uint64_t domain (extended witness set).
- Setup-side modular arithmetic (root/psi/q-1 factorization) avoids a
  compiler 128-bit type via a portable shift-and-add mulmod, keeping
  MSVC compatibility.
- Add Goldilocks (2^64 - 2^32 + 1) to the known q-1 factor table used
  for auto-deriving primitive roots.

+++ include/ntt/ntt.h
@@ -82,9 +82,9 @@ void ntt_destroy(ntt_ctx *ctx);
  * @return NTT_OK Multiplication completed successfully.
  * @return NTT_ERROR on errors.
  */
-int ntt_negacyclic_mul(uint32_t *a,
-                       uint32_t *b,
-                       uint32_t *c,
+int ntt_negacyclic_mul(uint64_t *a,
+                       uint64_t *b,
+                       uint64_t *c,
                        const ntt_ctx *ctx);
 
 /**
@@ -100,7 +100,7 @@ int ntt_negacyclic_mul(uint32_t *a,
  * @return NTT_OK  Transform completed successfully.
  * @return NTT_ERROR Invalid input.
  */
-int ntt_forward(const ntt_ctx *ctx, uint32_t *a);
+int ntt_forward(const ntt_ctx *ctx, uint64_t *a);
 
 /**
  * @brief Computes the inverse Number Theoretic Transform (INTT).
@@ -116,7 +116,7 @@ int ntt_forward(const ntt_ctx *ctx, uint32_t *a);
  * @return NTT_OK Transform completed successfully.
  * @return NTT_ERROR Invalid input.
  */
-int ntt_inverse(const ntt_ctx *ctx, uint32_t *a);
+int ntt_inverse(const ntt_ctx *ctx, uint64_t *a);
 
 #ifdef __cplusplus
 }
